Update Fte Formatting. Add Log in/Sign up page. Add Pet Lists

+++ src/components/SignUpForm/SignUpForm.jsx
@@ -1,45 +1,65 @@
 import React, { useState } from "react";
 import { Container, Form, Button } from "react-bootstrap";
+import { MDBInput} from "mdbreact";
+import "./SignUpForm.css";
 
-export default function SignUpForm() {
-    const [state, setState] = useState({
+export default function SignUpForm(props) {
+    const [signUpState, setSignUpState] = useState({
         name: "",
         email: "",
         password: "",
         phoneNumber: "",
         postalCode: "",
         error: "",
+        passwordConfirm: ""
     });
 
     const handleChange = (evt) => {
-        setState({ ...state, [evt.target.name]: evt.target.value, error: "" });
+        setSignUpState({ ...signUpState, [evt.target.name]: evt.target.value, error: "" });
     };
 
     const handleSubmit = async (evt) => {
         evt.preventDefault();
-        try {
-            const fetchResponse = await fetch("/api/users/signup", {
-                method: "POST",
-                headers: { "Content-Type": "application/json" },
-                body: JSON.stringify({
-                    name: state.name,
-                    email: state.email,
-                    password: state.password,
-                    phoneNumber: state.phoneNumber,
-                    postalCode: state.postalCode,
-                }),
-            });
-
-            if (!fetchResponse.ok) throw new Error("Fetch failed - Bad request");
-
-            let token = await fetchResponse.json();
-            localStorage.setItem("token", token);
-
-            const userDoc = JSON.parse(atob(token.split(".")[1])).user;
-            this.props.setUserInState(userDoc);
-        } catch (err) {
-            console.log("SignupForm error", err.message);
-            setState({ error: "Sign Up Failed - Try Again" });
+        if (signUpState.password !== signUpState.passwordConfirm ){
+            console.log("signupState.password",signUpState.password)
+            console.log("passwordState", signUpState.passwordConfirm )
+            alert("Passwords don't match")
+        } else {
+            try {
+                console.log(signUpState)
+                const fetchResponse = await fetch("/api/users/signup", {
+                    method: "POST",
+                    headers: { "Content-Type": "application/json" },
+                    body: JSON.stringify({
+                        name: signUpState.name,
+                        email: signUpState.email,
+                        password: signUpState.password,
+                        phoneNumber: signUpState.phoneNumber,
+                        postalCode: signUpState.postalCode,
+                    }),
+                });
+    
+                if (!fetchResponse.ok) throw new Error("Fetch failed - Bad request");
+    
+                let token = await fetchResponse.json();
+                localStorage.setItem("token", token);
+    
+                const userDoc = JSON.parse(atob(token.split(".")[1])).user;
+                props.setUserInState(userDoc);
+                setSignUpState({
+                    name: "",
+                    email: "",
+                    password: "",
+                    phoneNumber: "",
+                    postalCode: "",
+                    error: "",
+                    passwordConfirm: "",
+                })
+                
+            } catch (err) {
+                console.log("SignupForm error", err.message);
+                setSignUpState({ error: "Sign Up Failed - Try Again" });
+            }
         }
     };
 
@@ -47,65 +67,62 @@ export default function SignUpForm() {
         <div>
             <Container className="align-items-top justify-content-center d-flex">
                 <Form onSubmit={handleSubmit}>
-                    <Form.Group>
-                        <Form.Control
-                            name="name"
-                            value={state.name}
-                            onChange={handleChange}
-                            type="text"
-                            placeholder="Name"
-                            required
-                        />
-                    </Form.Group>
-                    <Form.Group>
-                        <Form.Control
-                            name="email"
-                            type="email"
-                            value={state.email}
-                            onChange={handleChange}
-                            required
-                            placeholder="Email"
-                        />
-                    </Form.Group>
-                    <Form.Group>
-                        <Form.Control
-                            name="password"
-                            type="password"
-                            value={state.password}
-                            onChange={handleChange}
-                            placeholder="Password"
-                            required
-                        />
-                    </Form.Group>
-                    <Form.Group>
-                        <Form.Control
-                            name="password2"
-                            type="password"
-                            placeholder="Password Confirmation"
-                            required
-                        />
-                    </Form.Group>
-                    <Form.Group>
-                        <Form.Control
-                            name="phoneNumber"
-                            type="tel"
-                            value={state.phoneNumber}
-                            onChange={handleChange}
-                            placeholder="Phone Number"
-                            required
-                        />
-                    </Form.Group>
-                    <Form.Group>
-                        <Form.Control
-                            name="postalCode"
-                            type="text"
-                            value={state.postalCode}
-                            onChange={handleChange}
-                            placeholder="Postal Code"
-                            required
-                        />
-                    </Form.Group>
-                    <Button type="submit" className="me-2 mt-3">
+                    <MDBInput
+                        label="Name"
+                        name="name"
+                        icon="user-alt"
+                        value={signUpState.name}
+                        onChange={handleChange}
+                        type="text"
+                        placeholder="Name"
+                        required
+                    />
+                    <MDBInput
+                        label="Email"
+                        icon="envelope"
+                        name="email"
+                        type="text"
+                        value={signUpState.email}
+                        onChange={handleChange}
+                        required
+                    />
+                    <MDBInput
+                        label="Password"
+                        icon="key"
+                        name="password"
+                        type="password"
+                        value={signUpState.password}
+                        onChange={handleChange}
+                        required
+                    />
+                     <MDBInput
+                        label="Password Confirmation"
+                        icon="key"
+                        name="passwordConfirm"
+                        type="password"
+                        value={signUpState.passwordConfirm}
+                        onChange={handleChange}
+                        required
+                    />
+                    <MDBInput
+                        label="Phone Number"
+                        icon="phone"
+                        name="phoneNumber"
+                        type="tel"
+                        value={signUpState.phoneNumber}
+                        onChange={handleChange}
+                        required
+                    />
+                    <MDBInput
+                        label="Postal Code"
+                        icon="map-marker-alt"
+                        name="postalCode"
+                        type="text"
+                        value={signUpState.postalCode}
+                        onChange={handleChange}
+                        required
+                    />
+                    <Button type="submit" className="SignUpButton me-2 mt-3">
                         Sign Up
                     </Button>
                 </Form>
